+++ resources/views/components/modal/save.blade.php
@@ -7,7 +7,7 @@
                 <div class="bx-title">
                     {{ $title }}
                 </div>
-                <x-gt-icon-cross wire:click="$toggle('showModal')" class="close sm" />
+                <x-gt-icon name="x-mark" wire:click="$toggle('showModal')" class="close sm" />
             </div>
         @endisset