+++ resources/views/components/v2/toast.blade.php
@@ -6,8 +6,8 @@
     x-transition.out.opacity.duration.1000ms>
 
     <div class="bx pxy-1 fixed pos-b pos-r mxy flex va-c minw-16 z-top">
-        <x-gt-icon-tick-round class="txt-green" />
+        <x-gt-icon name="check-circle" class="txt-green" />
         <div x-text="message" class="mx-1"></div>
-        <x-gt-icon-cross class="fs0 close icon ml-auto" @click="show = false" />
+        <x-gt-icon name="x-mark" class="fs0 close ml-auto" @click="show = false" />
     </div>
 </div>